requirements.txt
accelerate==1.12.0
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
attrs==25.4.0
av==16.1.0
bcrypt==4.1.2
bitsandbytes==0.49.1
boto3==1.40.61
botocore==1.40.61
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
cryptography==46.0.3
ctranslate2==4.6.3
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.128.0
faster-whisper==1.2.1
ffmpeg-python==0.2.0
filelock==3.20.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
greenlet==3.3.0
h11==0.16.0
hf-xet==1.2.0
httptools==0.7.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.10
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.23.2
packaging==25.0
passlib==1.7.4
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.21
PyYAML==6.0.3
qwen-vl-utils==0.0.14
redis==7.1.0
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
safetensors==0.7.0
six==1.17.0
SQLAlchemy==2.0.27
starlette==0.50.0
sympy==1.13.1
tokenizers==0.22.2
tomli==2.4.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.57.6
triton==3.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
wrapt==1.17.3
yarl==1.22.0
yt-dlp==2025.12.8
